1.messageId扩展到64位 2.m2m持久化方式修改 3.被挤下线使用DisconnectMessage替换直接close掉session

+++ broker/src/main/java/org/eclipse/moquette/server/netty/NettyMQTTHandler.java
@@ -78,11 +78,13 @@ public class NettyMQTTHandler extends ChannelInboundHandlerAdapter {
     @Override
     public void channelInactive(ChannelHandlerContext ctx) throws Exception {
         NettyChannel channel = m_channelMapper.get(ctx);
-        String clientID = (String) channel.getAttribute(Constants.ATTR_CLIENTID);
-        m_messaging.lostConnection(channel, clientID);
-        ctx.close(/*false*/);
-        synchronized(m_channelMapper) {
-            m_channelMapper.remove(ctx);
+        if (channel != null) {
+            String clientID = (String) channel.getAttribute(Constants.ATTR_CLIENTID);
+            m_messaging.lostConnection(channel, clientID);
+            ctx.close(/*false*/);
+            synchronized(m_channelMapper) {
+                m_channelMapper.remove(ctx);
+            }
         }
     }
